Move Mesh_3 specific properties to Mesh_3
