+++ Mesh_3/examples/Mesh_3/mesh_polyhedral_domain_with_two_polyhedra.cpp
@@ -6,7 +6,7 @@
 
 #include <CGAL/Surface_mesh.h>
 #include <CGAL/boost/graph/Graph_with_descriptor_with_graph.h>
-#include <CGAL/boost/graph/PMP_properties_Surface_mesh.h>
+#include <CGAL/Mesh_3/properties_Surface_mesh.h>
 #include <CGAL/Polyhedral_mesh_domain_with_features_3.h>
 #include <CGAL/make_mesh_3.h>
 #include <CGAL/Timer.h>
