+++ Mesh_3/include/CGAL/Mesh_3/properties.h
